refactor: keep 'platformDeveloperDocs' as color-preset name for tutorials
