+++ packages/gatsby-theme-docs/color-presets/platform-developer-docs/index.js
